+++ Pac-Man-DeLuxe/Pac-Man-DeLuxe/Vector2.cpp
@@ -5,7 +5,7 @@ Vector2::Vector2() {
 	this->y_ = 0;
 }
 
-Vector2::Vector2(float x, float y) {
+Vector2::Vector2(int x, int y) {
 	this->x_ = x;
 	this->y_ = y;
 }
